#151: move ballotPapersPathObject to correct folder
