+++ packages/votura-validators/src/schema/ballotPaperSections/ballotPaperSectionsPathObject.ts
@@ -3,6 +3,7 @@ import {
   insertableBallotPaperSectionObjectSchema,
   selectableBallotPaperSectionObjectSchema,
 } from '../../objects/ballotPaperSection.js';
+import { ballotPaperIdParameter } from '../ballotPapers/ballotPaperIdParameter.js';
 import { electionIdParameter } from '../elections/electionIdParameter.js';
 import {
   response400,
@@ -16,7 +17,6 @@ import {
 } from '../globals/responses.js';
 import { SecuritySchemaName } from '../globals/securitySchemaName.js';
 import { Tag } from '../globals/tag.js';
-import { ballotPaperIdParameter } from '../ballotPapers/ballotPaperIdParameter.js';
 
 export const ballotPaperSectionsPathObject: OpenAPIV3.PathItemObject = {
   summary: 'Get and create ballot paper sections',
