+++ packages/votura-validators/src/schema/voturaOpenApiSchema.ts
@@ -1,10 +1,10 @@
 import type { OpenAPIV3 } from 'openapi-types';
 import { ballotPaperIdPathObject } from './ballotPapers/ballotPaperIdPathObject.js';
-import { ballotPaperSectionsPathObject } from './ballotPaperSections/ballotPaperSectionsPathObject.js';
+import { ballotPapersPathObject } from './ballotPapers/ballotPapersPathObject.js';
 import { ballotPaperSectionIdPathObject } from './ballotPaperSections/ballotPaperSectionIdPathObject.js';
+import { ballotPaperSectionsPathObject } from './ballotPaperSections/ballotPaperSectionsPathObject.js';
 import { candidateIdPathObject } from './candidates/candidateIdPathObject.js';
 import { candidatesPathObject } from './candidates/candidatesPathObject.js';
-import { ballotPapersPathObject } from './elections/ballotPapersPathObject.js';
 import { electionIdPathObject } from './elections/electionIdPathObject.js';
 import { electionsPathObject } from './elections/electionsPathObject.js';
 import { freezePathObject } from './elections/freezePathObject.js';
